pyproject.toml
[project]
name = "portfolio-sentiment-agent"
version = "0.1.0"
description = "Multi-agent ML system for portfolio sentiment analysis"
requires-python = ">=3.11"
dependencies = [
    "python-dotenv>=1.0.0",
    "pydantic>=2.5.0",
    "sqlalchemy>=2.0.23",
    "psycopg2-binary>=2.9.9",
    "torch>=2.1.0",
    "transformers>=4.35.0",
    "sentencepiece>=0.1.99",
    "requests>=2.31.0",
    "sendgrid>=6.11.0",
    "anthropic>=0.18.0",
    "openai>=1.6.0",
    "langgraph>=0.0.20",
    "langchain>=0.1.0",
    "langchain-core>=0.1.0",
    "pyyaml>=6.0.1",
    "python-dateutil>=2.8.2",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.3",
    "pytest-cov>=4.1.0",
    "ruff>=0.1.8",
    "black>=23.12.0",
    "mypy>=1.7.0",
]

[tool.black]
line-length = 100
target-version = ['py311']

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "--cov=agents --cov=services --cov=db --cov=config --cov-report=term-missing --cov-report=html"

